backend/requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
psycopg2-binary==2.9.7
pytest==7.4.0
httpx==0.23.3
